Scrimmage v0.5

+++ src/Atlas.cpp
@@ -26,7 +26,7 @@ void Atlas::Clear()
 }
 int Atlas::GetSize()
 {
-	return (int)imgList.size();
+	return imgList.size();
 }
 IMAGE* Atlas::GetImage(int idx)
 {
