+++ src/Util.h
@@ -14,7 +14,7 @@ public:
 inline void DebugLine(Camera& camera, int x1, int y1, int x2, int y2)
 {
 	Vector2 pos_camera = camera.GetPos();
-	line((int)(x1 - pos_camera.x), (int)(y1 - pos_camera.y), (int)(x2 - pos_camera.x), (int)(y2 - pos_camera.y));
+	line(x1 - pos_camera.x, y1 - pos_camera.y, x2 - pos_camera.x, y2 - pos_camera.y);
 }
 
 inline void FlipImage(IMAGE* src, IMAGE* dst)
@@ -47,7 +47,7 @@ inline void PutImageAlpha(Camera& camera, int dstx, int dsty, IMAGE* img)
 {
 	int w = img->getwidth();
 	int h = img->getheight();
-	AlphaBlend(GetImageHDC(GetWorkingImage()), (int)(dstx - camera.GetPos().x), (int)(dsty - camera.GetPos().y), w, h,
+	AlphaBlend(GetImageHDC(GetWorkingImage()), (dstx - camera.GetPos().x), (dsty - camera.GetPos().y), w, h,
 		GetImageHDC(img), 0, 0, w, h, { AC_SRC_OVER, 0, 255, AC_SRC_ALPHA });
 }
 inline void PutImageAlpha(int dstx, int dsty, int width, int height, IMAGE* img, int srcx, int srcy)
