+++ src/Vector2.h
@@ -7,22 +7,22 @@ class Vector2
 public:
 	Vector2();
 	~Vector2();
-	Vector2(float x, float y);
+	Vector2(int x, int y);
 private:
 	
 public:
-	float x;
-	float y;
+	int x;
+	int y;
 public:
 	void Init();
-	float Length() const;
+	int Length() const;
 	Vector2 Normalize() const;
 
 	Vector2 V2PlusVec(Vector2& vec);
 	void PlusVec(Vector2& vec);
 	Vector2 V2MinusVec(Vector2& vec);
 	void MinusVec(Vector2& vec);
-	Vector2 V2MultiVal(float& val);
-	float SumMultiVec(Vector2& vec);
-	void MultiVal(float& val);
+	Vector2 V2MultiVal(int& val);
+	int SumMultiVec(Vector2& vec);
+	void MultiVal(int& val);
 };
